Generalise the existing service
